Refactor into ParserUtil
