+++ src/main/java/seedu/address/logic/parser/ExportCommandParser.java
@@ -3,13 +3,13 @@ package seedu.address.logic.parser;
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ATTRIBUTE;
+import static seedu.address.logic.parser.ParserUtil.parseAttributes;
 
-import java.util.EnumSet;
 import java.util.List;
 
 import seedu.address.logic.commands.ExportCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.Attribute;
+import seedu.address.model.person.Attribute;
 
 /**
  * Parses input arguments and creates a new ExportCommand object
@@ -24,24 +24,13 @@ public class ExportCommandParser implements Parser<ExportCommand> {
         requireNonNull(args);
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_ATTRIBUTE);
-        String errorMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT, ExportCommand.MESSAGE_USAGE);
+        // Since `export` is a valid command without prefixes, we ignore the error message from parseAttributes
         try {
-            List<Attribute> attributes = argMultimap.getAllValues(PREFIX_ATTRIBUTE).stream()
-                    .map(String::toUpperCase)
-                    .map(Attribute::valueOf)
-                    .toList();
-
-            // Check if there are duplicates
-            if (!attributes.isEmpty()) {
-                // copyOf could throw if we don't make sure attributes is non-empty
-                EnumSet<Attribute> uniqueValues = EnumSet.copyOf(attributes);
-                if (uniqueValues.size() != attributes.size()) {
-                    throw new ParseException(errorMessage);
-                }
-            }
+            List<Attribute> attributes = parseAttributes(argMultimap.getAllValues(PREFIX_ATTRIBUTE));
             return new ExportCommand(attributes);
-        } catch (IllegalArgumentException e) {
-            throw new ParseException(errorMessage);
+        } catch (ParseException e) {
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, ExportCommand.MESSAGE_USAGE));
         }
+
     }
 }
