+++ src/main/java/seedu/address/logic/parser/ParserUtil.java
@@ -2,14 +2,18 @@ package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
 
+import java.util.Arrays;
 import java.util.Collection;
+import java.util.EnumSet;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import seedu.address.commons.core.index.Index;
 import seedu.address.commons.util.StringUtil;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.person.Address;
+import seedu.address.model.person.Attribute;
 import seedu.address.model.person.Commission;
 import seedu.address.model.person.Email;
 import seedu.address.model.person.Name;
@@ -122,6 +126,31 @@ public class ParserUtil {
             tagSet.add(parseTag(tagName));
         }
         return tagSet;
+    }
+
+    /**
+     * Parses {@code Collection<String> attributes} into a {@code List<Attribute>}.
+     */
+    public static List<Attribute> parseAttributes(Collection<String> attributes) throws ParseException {
+        requireNonNull(attributes);
+        if (attributes.isEmpty()) {
+            return Arrays.asList(Attribute.values());
+        }
+        try {
+            List<Attribute> result = attributes.stream()
+                    .map(String::trim)
+                    .map(String::toUpperCase)
+                    .map(Attribute::valueOf)
+                    .toList();
+            // copyOf could throw if we don't make sure attributes is non-empty
+            EnumSet<Attribute> uniqueValues = EnumSet.copyOf(result);
+            if (uniqueValues.size() != result.size()) {
+                throw new ParseException(Attribute.MESSAGE_CONSTRAINTS);
+            }
+            return result;
+        } catch (IllegalArgumentException e) {
+            throw new ParseException(Attribute.MESSAGE_CONSTRAINTS);
+        }
     }
 
     /**
