+++ src/main/java/seedu/address/model/person/Attribute.java
@@ -1,4 +1,4 @@
-package seedu.address.model;
+package seedu.address.model.person;
 
 /**
  * Stores the details to edit the person with. Each non-empty field value will replace the
@@ -9,5 +9,8 @@ public enum Attribute {
     PHONE,
     EMAIL,
     ADDRESS,
-    COMMISSION,
+    COMMISSION;
+
+    public static final String MESSAGE_CONSTRAINTS = "ATTRIBUTE may be: "
+        + "name, phone, email, address, or commission, without duplicates";
 }
