+++ src/test/java/seedu/address/model/util/ExportDataUtilTest.java
@@ -15,10 +15,10 @@ import java.util.stream.Stream;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import seedu.address.model.Attribute;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
+import seedu.address.model.person.Attribute;
 
 public class ExportDataUtilTest {
